pyproject.toml
[project]
name = "autotest-frame"
version = "0.1.0"
description = "A modular and extensible Python automation testing framework based on pytest"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "adbutil>=0.0.6",
    "allure-pytest>=2.15.0",
    "colorlog>=6.9.0",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pytest>=8.4.2",
    "pytest-assume>=2.4.3",
    "pytest-rerunfailures>=16.0.1",
    "pyyaml>=6.0.2",
    "scikit-image>=0.25.2",
    "setuptools>=80.9.0",
    "uiautomator2>=3.4.1",
    "weditor==0.6.4",
]
